+++ src/main/java/flab/payment_system/common/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package flab.payment_system.core.exception;
+package flab.payment_system.common.exception;
 
 import org.springframework.http.HttpStatus;
