+++ src/main/java/flab/payment_system/common/interceptor/LoggingInterceptor.java
@@ -1,8 +1,7 @@
-package flab.payment_system.core.interceptor;
+package flab.payment_system.common.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.checkerframework.checker.nullness.qual.NonNull;
