+++ src/main/java/flab/payment_system/common/response/ResponseMessage.java
@@ -1,4 +1,4 @@
-package flab.payment_system.core.response;
+package flab.payment_system.common.response;
 
 public record ResponseMessage(String message) {
