+++ src/main/java/flab/payment_system/config/AppConfig.java
@@ -1,9 +1,9 @@
 package flab.payment_system.config;
 
 
-import flab.payment_system.core.enums.Constant;
-import flab.payment_system.core.filter.ExceptionHandlerFilter;
-import flab.payment_system.core.filter.SignInCheckFilter;
+import flab.payment_system.common.enums.Constant;
+import flab.payment_system.common.filter.ExceptionHandlerFilter;
+import flab.payment_system.common.filter.SignInCheckFilter;
 import flab.payment_system.domain.session.service.SessionService;
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
