+++ src/main/java/flab/payment_system/domain/order/domain/OrderProduct.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.order.domain;
 
-import flab.payment_system.core.data.BaseEntity;
+import flab.payment_system.common.data.BaseEntity;
 import jakarta.annotation.Nonnull;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
