+++ src/main/java/flab/payment_system/domain/payment/exception/PaymentNotExistBadRequestException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.payment.exception;
 
-import flab.payment_system.core.exception.BadRequestException;
+import flab.payment_system.common.exception.BadRequestException;
 
 public class PaymentNotExistBadRequestException extends BadRequestException {
 	public PaymentNotExistBadRequestException() {
