+++ src/main/java/flab/payment_system/domain/payment/exception/PaymentTossServiceUnavailableException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.payment.exception;
 
-import flab.payment_system.core.exception.ServiceUnavailableException;
+import flab.payment_system.common.exception.ServiceUnavailableException;
 
 public class PaymentTossServiceUnavailableException extends ServiceUnavailableException {
