+++ src/main/java/flab/payment_system/domain/product/domain/Product.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.product.domain;
 
-import flab.payment_system.core.data.BaseEntity;
+import flab.payment_system.common.data.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
