+++ src/main/java/flab/payment_system/domain/product/exception/ProductNotExistBadRequestException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.product.exception;
 
-import flab.payment_system.core.exception.BadRequestException;
+import flab.payment_system.common.exception.BadRequestException;
 
 public class ProductNotExistBadRequestException extends BadRequestException {
