+++ src/main/java/flab/payment_system/domain/product/exception/ProductSoldOutException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.product.exception;
 
-import flab.payment_system.core.exception.OkException;
+import flab.payment_system.common.exception.OkException;
 
 public class ProductSoldOutException extends OkException {
