+++ src/main/java/flab/payment_system/domain/user/exception/UserEmailAlreadyExistConflictException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.user.exception;
 
-import flab.payment_system.core.exception.ConflictException;
+import flab.payment_system.common.exception.ConflictException;
 
 public class UserEmailAlreadyExistConflictException extends ConflictException {
