+++ src/main/java/flab/payment_system/domain/user/exception/UserVerificationUnauthorizedException.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.user.exception;
 
-import flab.payment_system.core.exception.UnauthorizedException;
+import flab.payment_system.common.exception.UnauthorizedException;
 
 public class UserVerificationUnauthorizedException extends UnauthorizedException {
